@@ -24,11 +24,11 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
   json_annotation: ^2.0.0
-  cached_network_image: ^1.1.2+1
-  graphql_flutter: ^1.0.0-beta
-  flutter_bloc: ^0.11.0
-  equatable: ^0.2.0
-  shared_preferences: ^0.5.2
+  cached_network_image: ^2.0.0
+  graphql_flutter: ^3.0.0
+  flutter_bloc: ^3.2.0
+  equatable: ^1.1.1
+  shared_preferences: ^0.5.6+3
 
 dev_dependencies:
   flutter_test:
