@@ -7,26 +7,25 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 
-class SimpleBlocDelegate extends BlocDelegate {
-  @override
-  onTransition(Transition transition) {
-    super.onTransition(transition);
-    print(transition);
-  }
-}
-
 void main() {
-  BlocSupervisor().delegate = SimpleBlocDelegate();
+  BlocSupervisor.delegate = SimpleBlocDelegate();
 
   final HttpLink _httpLink = HttpLink(uri: TVDB_GRAPHQL_API);
 
   final GraphQLClient _client =
-      GraphQLClient(link: _httpLink as Link, cache: InMemoryCache());
+      GraphQLClient(link: _httpLink, cache: InMemoryCache());
 
   final TvdbRepository tvdbRepository =
       TvdbRepository(tvdbGraphQLClient: TvdbGraphQLClient(client: _client));
 
-  return runApp(MyApp(tvdbRepository: tvdbRepository));
+  return runApp(
+    BlocProvider(
+      create: (context) {
+        return FavoritesBloc()..add(FetchFavorites());
+      },
+      child: MyApp(tvdbRepository: tvdbRepository),
+    ),
+  );
 }
 
 class MyApp extends StatefulWidget {
@@ -39,74 +38,55 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  final FavoritesBloc favoritesBloc = FavoritesBloc();
-  NextEpisodesBloc _nextEpisodesBloc;
-  SeriesDetailsBloc _seriesDetailsBloc;
-  SearchSeriesBloc _searchSeriesBloc;
-
-  @override
-  void initState() {
-    super.initState();
-    _nextEpisodesBloc = NextEpisodesBloc(
-        tvdbRepository: widget.tvdbRepository, favoritesBloc: favoritesBloc);
-    _seriesDetailsBloc =
-        SeriesDetailsBloc(tvdbRepository: widget.tvdbRepository);
-    _searchSeriesBloc = SearchSeriesBloc(tvdbRepository: widget.tvdbRepository);
-  }
-
   @override
   Widget build(BuildContext context) {
-    return BlocProvider(
-      bloc: _nextEpisodesBloc,
-      child: BlocProviderTree(
-        blocProviders: [
-          BlocProvider<FavoritesBloc>(bloc: favoritesBloc),
-          BlocProvider<NextEpisodesBloc>(bloc: _nextEpisodesBloc),
-          BlocProvider<SeriesDetailsBloc>(bloc: _seriesDetailsBloc),
-          BlocProvider<SearchSeriesBloc>(bloc: _searchSeriesBloc),
-        ],
-        child: MaterialApp(
-          title: 'Episode Guide',
-          debugShowCheckedModeBanner: false,
-          theme: ThemeData(
-            primaryColorDark: Colors.grey[900],
-            primaryColorLight: Colors.grey[500],
-            primaryColor: Colors.black,
-            accentColor: Colors.white,
-            primarySwatch: Colors.blue,
-            fontFamily: 'AlegreyaSans',
-            textTheme: TextTheme(
-              headline: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-              subhead: TextStyle(
-                fontSize: 18,
-                color: Colors.white,
-              ),
+    return MultiBlocProvider(
+      providers: [
+        BlocProvider<NextEpisodesBloc>(
+          create: (context) => NextEpisodesBloc(
+              favoritesBloc: BlocProvider.of<FavoritesBloc>(context),
+              tvdbRepository: widget.tvdbRepository),
+        ),
+        BlocProvider<SeriesDetailsBloc>(
+          create: (context) =>
+              SeriesDetailsBloc(tvdbRepository: widget.tvdbRepository),
+        ),
+        BlocProvider<SearchSeriesBloc>(
+          create: (context) =>
+              SearchSeriesBloc(tvdbRepository: widget.tvdbRepository),
+        ),
+      ],
+      child: MaterialApp(
+        title: 'Episode Guide',
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          primaryColorDark: Colors.grey[900],
+          primaryColorLight: Colors.grey[500],
+          primaryColor: Colors.black,
+          accentColor: Colors.white,
+          primarySwatch: Colors.blue,
+          fontFamily: 'AlegreyaSans',
+          textTheme: TextTheme(
+            headline5: TextStyle(
+              fontSize: 24,
+              fontWeight: FontWeight.bold,
+              color: Colors.white,
+            ),
+            subtitle1: TextStyle(
+              fontSize: 18,
+              color: Colors.white,
             ),
           ),
-          initialRoute: HomePage.routeName,
-          routes: {
-            HomePage.routeName: (context) {
-              return HomePage(
-                onInit: () => favoritesBloc.dispatch(FetchFavorites()),
-              );
-            },
-            SeriesDetailsScreen.routeName: (context) => SeriesDetailsScreen(),
-            SearchSeriesScreen.routeName: (context) => SearchSeriesScreen(),
-          },
         ),
+        initialRoute: HomePage.routeName,
+        routes: {
+          HomePage.routeName: (context) => HomePage(
+                onInit: () {},
+              ),
+          SeriesDetailsScreen.routeName: (context) => SeriesDetailsScreen(),
+          SearchSeriesScreen.routeName: (context) => SearchSeriesScreen(),
+        },
       ),
     );
-  }
-
-  @override
-  void dispose() {
-    _seriesDetailsBloc.dispose();
-    _nextEpisodesBloc.dispose();
-    _searchSeriesBloc.dispose();
-    super.dispose();
   }
 }
