@@ -1,6 +1,5 @@
-export './home/carousel.dart';
 export './home/episode_card.dart';
 export './home/home_page.dart';
-export './series/series_details.dart';
 export './search/search_page.dart';
 export './search/search_result_card.dart';
+export './series/series_details.dart';
